requirements.txt
grpcio
grpcio-tools
PyYAML
sphinx
cloud_sptheme
